updated test cases, added minor docs for iterator implementation
