+++ test_dir/custom.ignore
@@ -1,1 +1,1 @@
-turkey.file
+chicken/
